Remove upgrade-legacy-config as a top-level task
Instead, add a "--legay" flag to the "upgrade-config" task. Also
updates callers to use this new method of invocation.
