+++ framework/builtintasks/upgradeconfig.go
@@ -22,11 +22,15 @@ import (
 	"os"
 	"path"
 	"path/filepath"
+	"sort"
 	"strings"
 
+	"github.com/palantir/pkg/matcher"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
+	"gopkg.in/yaml.v2"
 
+	"github.com/palantir/godel/framework/godel/config"
 	"github.com/palantir/godel/framework/godellauncher"
 )
 
@@ -35,11 +39,13 @@ func UpgradeConfigTask(upgradeTasks []godellauncher.UpgradeConfigTask) godellaun
 		upgradeConfigCmdName = "upgrade-config"
 		dryRunFlagName       = "dry-run"
 		printContentFlagName = "print-content"
+		legacyFlagName       = "legacy"
 	)
 
 	var (
 		dryRunFlagVal       bool
 		printContentFlagVal bool
+		legacyFlagVal       bool
 	)
 
 	cmd := &cobra.Command{
@@ -49,6 +55,7 @@ func UpgradeConfigTask(upgradeTasks []godellauncher.UpgradeConfigTask) godellaun
 
 	cmd.Flags().BoolVar(&dryRunFlagVal, dryRunFlagName, false, "print what the upgrade operation would do without writing changes")
 	cmd.Flags().BoolVar(&printContentFlagVal, printContentFlagName, false, "print the content of the changes to stdout in addition to writing them")
+	cmd.Flags().BoolVar(&legacyFlagVal, legacyFlagName, false, "upgrade pre-2.0 legacy configuration")
 
 	cmd.SilenceErrors = true
 	cmd.SilenceUsage = true
@@ -70,28 +77,10 @@ func UpgradeConfigTask(upgradeTasks []godellauncher.UpgradeConfigTask) godellaun
 				if err != nil {
 					return err
 				}
-
-				var failedUpgrades []string
-				for _, upgradeTask := range upgradeTasks {
-					changed, upgradedCfgBytes, err := upgradeConfigFile(upgradeTask, global, configDirPath, dryRunFlagVal, stdout)
-					if err != nil {
-						failedUpgrades = append(failedUpgrades, upgradeError(projectDir, path.Join(configDirPath, upgradeTask.ConfigFile), err))
-						continue
-					}
-					if !changed {
-						continue
-					}
-					printUpgradedConfig(upgradeTask.ConfigFile, upgradedCfgBytes, dryRunFlagVal, printContentFlagVal, stdout)
+				if legacyFlagVal {
+					return runUpgradeLegacyConfig(upgradeTasks, global, projectDir, configDirPath, dryRunFlagVal, printContentFlagVal, cmd.OutOrStdout())
 				}
-
-				if len(failedUpgrades) == 0 {
-					return nil
-				}
-				dryRunPrintln(stdout, dryRunFlagVal, "Failed to upgrade configuration:")
-				for _, upgrade := range failedUpgrades {
-					dryRunPrintln(stdout, dryRunFlagVal, "\t"+upgrade)
-				}
-				return fmt.Errorf("")
+				return runUpgradeConfig(upgradeTasks, global, projectDir, configDirPath, dryRunFlagVal, printContentFlagVal, cmd.OutOrStdout())
 			}
 
 			rootCmd := godellauncher.CobraCmdToRootCmd(cmd)
@@ -99,6 +88,37 @@ func UpgradeConfigTask(upgradeTasks []godellauncher.UpgradeConfigTask) godellaun
 			return rootCmd.Execute()
 		},
 	}
+}
+
+func runUpgradeConfig(
+	upgradeTasks []godellauncher.UpgradeConfigTask,
+	global godellauncher.GlobalConfig,
+	projectDir, configDirPath string,
+	dryRun, printContent bool,
+	stdout io.Writer,
+) error {
+
+	var failedUpgrades []string
+	for _, upgradeTask := range upgradeTasks {
+		changed, upgradedCfgBytes, err := upgradeConfigFile(upgradeTask, global, configDirPath, dryRun, stdout)
+		if err != nil {
+			failedUpgrades = append(failedUpgrades, upgradeError(projectDir, path.Join(configDirPath, upgradeTask.ConfigFile), err))
+			continue
+		}
+		if !changed {
+			continue
+		}
+		printUpgradedConfig(upgradeTask.ConfigFile, upgradedCfgBytes, dryRun, printContent, stdout)
+	}
+
+	if len(failedUpgrades) == 0 {
+		return nil
+	}
+	dryRunPrintln(stdout, dryRun, "Failed to upgrade configuration:")
+	for _, upgrade := range failedUpgrades {
+		dryRunPrintln(stdout, dryRun, "\t"+upgrade)
+	}
+	return fmt.Errorf("")
 }
 
 func dryRunPrintln(w io.Writer, dryRun bool, content string) {
@@ -186,4 +206,276 @@ func upgradeError(projectDir, configFilePath string, upgradeErr error) string {
 		}
 	}
 	return fmt.Sprintf("%s: %v", configFilePath, upgradeErr)
+}
+
+func runUpgradeLegacyConfig(
+	upgradeTasks []godellauncher.UpgradeConfigTask,
+	global godellauncher.GlobalConfig,
+	projectDir, configDirPath string,
+	dryRun, printContent bool,
+	stdout io.Writer,
+) error {
+
+	// record all of the original YML files in the directory
+	originalYMLFiles, err := dirYMLFiles(configDirPath)
+	if err != nil {
+		return err
+	}
+	// track all of the upgraded YML files
+	knownConfigFiles := make(map[string]struct{})
+
+	upgradeTasksMap := make(map[string]godellauncher.UpgradeConfigTask)
+	for _, upgradeTask := range upgradeTasks {
+		upgradeTasksMap[upgradeTask.ID] = upgradeTask
+	}
+
+	var failedUpgrades []string
+	// perform hard-coded one-time upgrades
+	for _, currUpgrader := range hardCodedLegacyUpgraders {
+		if err := currUpgrader.upgradeConfig(configDirPath, dryRun, printContent, stdout); err != nil {
+			failedUpgrades = append(failedUpgrades, upgradeError(projectDir, path.Join(configDirPath, currUpgrader.configFileName()), err))
+		}
+		knownConfigFiles[currUpgrader.configFileName()] = struct{}{}
+	}
+
+	var legacyConfigUpgraderIDs []string
+	for _, upgradeTask := range upgradeTasks {
+		// consider current configuration file for the plugin as known (don't warn if these files already
+		// existed in config directory but were not processed by a legacy config upgrader).
+		knownConfigFiles[upgradeTask.ConfigFile] = struct{}{}
+		if upgradeTask.LegacyConfigFile == "" {
+			continue
+		}
+		legacyConfigUpgraderIDs = append(legacyConfigUpgraderIDs, upgradeTask.ID)
+	}
+	sort.Strings(legacyConfigUpgraderIDs)
+	for _, k := range legacyConfigUpgraderIDs {
+		upgradeTask, ok := upgradeTasksMap[k]
+		if !ok {
+			// legacy task does not have an upgrader: continue
+			continue
+		}
+		knownConfigFiles[upgradeTask.LegacyConfigFile] = struct{}{}
+		if err := upgradeLegacyConfig(upgradeTask, configDirPath, global, dryRun, printContent, stdout); err != nil {
+			failedUpgrades = append(failedUpgrades, upgradeError(projectDir, path.Join(configDirPath, upgradeTask.ConfigFile), err))
+			continue
+		}
+	}
+
+	var unhandledYMLFiles []string
+	for _, k := range originalYMLFiles {
+		if _, ok := knownConfigFiles[k]; ok {
+			continue
+		}
+		unhandledYMLFiles = append(unhandledYMLFiles, k)
+	}
+	if err := processUnhandledYMLFiles(configDirPath, unhandledYMLFiles, dryRun, stdout); err != nil {
+		return err
+	}
+
+	if len(failedUpgrades) == 0 {
+		return nil
+	}
+	dryRunPrintln(stdout, dryRun, "Failed to upgrade configuration:")
+	for _, upgrade := range failedUpgrades {
+		dryRunPrintln(stdout, dryRun, "\t"+upgrade)
+	}
+	return fmt.Errorf("")
+}
+
+var hardCodedLegacyUpgraders = []hardCodedLegacyUpgrader{
+	&hardCodedLegacyUpgraderImpl{
+		fileName: "exclude.yml",
+		upgradeConfigFn: func(configDirPath string, dryRun, printContent bool, stdout io.Writer) error {
+			// godel.yml itself is compatible. Only work to be performed is if "exclude.yml" exists and contains entries that
+			// differ from godel.yml.
+			legacyExcludeFilePath := path.Join(configDirPath, "exclude.yml")
+			if _, err := os.Stat(legacyExcludeFilePath); os.IsNotExist(err) {
+				// if legacy file does not exist, there is no upgrade to be performed
+				return nil
+			}
+			legacyConfigBytes, err := ioutil.ReadFile(legacyExcludeFilePath)
+			if err != nil {
+				return errors.Wrapf(err, "failed to read legacy configuration file")
+			}
+			var excludeCfg matcher.NamesPathsCfg
+			if err := yaml.UnmarshalStrict(legacyConfigBytes, &excludeCfg); err != nil {
+				return errors.Wrapf(err, "failed to unmarshal legacy exclude configuration")
+			}
+
+			currentGodelConfig, err := config.ReadGodelConfigFromFile(path.Join(configDirPath, "godel.yml"))
+			if err != nil {
+				return errors.Wrapf(err, "failed to read godel configuration")
+			}
+
+			existingNames := make(map[string]struct{})
+			for _, name := range currentGodelConfig.Exclude.Names {
+				existingNames[name] = struct{}{}
+			}
+			existingPaths := make(map[string]struct{})
+			for _, path := range currentGodelConfig.Exclude.Paths {
+				existingPaths[path] = struct{}{}
+			}
+
+			modified := false
+			for _, legacyName := range excludeCfg.Names {
+				if _, ok := existingNames[legacyName]; ok {
+					continue
+				}
+				currentGodelConfig.Exclude.Names = append(currentGodelConfig.Exclude.Names, legacyName)
+				modified = true
+			}
+			for _, legacyPath := range excludeCfg.Paths {
+				if _, ok := existingPaths[legacyPath]; ok {
+					continue
+				}
+				currentGodelConfig.Exclude.Names = append(currentGodelConfig.Exclude.Paths, legacyPath)
+				modified = true
+			}
+
+			// back up old configuration by moving it
+			if err := backupConfigFile(legacyExcludeFilePath, dryRun, stdout); err != nil {
+				return errors.Wrapf(err, "failed to back up legacy configuration file")
+			}
+
+			if !modified {
+				// exclude.yml did not provide any new excludes: no need to write
+				return nil
+			}
+
+			upgradedCfgBytes, err := yaml.Marshal(currentGodelConfig)
+			if err != nil {
+				return errors.Wrapf(err, "failed to marshal upgraded godel configuration")
+			}
+
+			if !dryRun {
+				// write migrated configuration
+				if err := ioutil.WriteFile(path.Join(configDirPath, "godel.yml"), upgradedCfgBytes, 0644); err != nil {
+					return errors.Wrapf(err, "failed to write upgraded configuration")
+				}
+			}
+			printUpgradedConfig("godel.yml", upgradedCfgBytes, dryRun, printContent, stdout)
+			return nil
+		},
+	},
+}
+
+func dirYMLFiles(inputDir string) ([]string, error) {
+	fis, err := ioutil.ReadDir(inputDir)
+	if err != nil {
+		return nil, errors.Wrapf(err, "failed to read input directory")
+	}
+	var ymlFiles []string
+	for _, fi := range fis {
+		if fi.IsDir() {
+			continue
+		}
+		if strings.HasSuffix(fi.Name(), ".yml") {
+			ymlFiles = append(ymlFiles, fi.Name())
+		}
+	}
+	return ymlFiles, nil
+}
+
+type hardCodedLegacyUpgrader interface {
+	configFileName() string
+	upgradeConfig(configDirPath string, dryRun, printContent bool, stdout io.Writer) error
+}
+
+type hardCodedLegacyUpgraderImpl struct {
+	fileName        string
+	upgradeConfigFn func(configDirPath string, dryRun, printContent bool, stdout io.Writer) error
+}
+
+func (u *hardCodedLegacyUpgraderImpl) configFileName() string {
+	return u.fileName
+}
+
+func (u *hardCodedLegacyUpgraderImpl) upgradeConfig(configDirPath string, dryRun, printContent bool, stdout io.Writer) error {
+	return u.upgradeConfigFn(configDirPath, dryRun, printContent, stdout)
+}
+
+func upgradeLegacyConfig(upgradeTask godellauncher.UpgradeConfigTask, configDirPath string, global godellauncher.GlobalConfig, dryRun, printContent bool, stdout io.Writer) error {
+	legacyConfigFilePath := path.Join(configDirPath, upgradeTask.LegacyConfigFile)
+	if _, err := os.Stat(legacyConfigFilePath); os.IsNotExist(err) {
+		// if legacy file does not exist, there is no upgrade to be performed
+		return nil
+	}
+	legacyConfigBytes, err := ioutil.ReadFile(legacyConfigFilePath)
+	if err != nil {
+		return errors.Wrapf(err, "failed to read legacy configuration file")
+	}
+
+	var ymlConfig yaml.MapSlice
+	if err := yaml.Unmarshal(legacyConfigBytes, &ymlConfig); err != nil {
+		return errors.Wrapf(err, "failed to unmarshal YAML configuration")
+	}
+	// add "legacy-config: true" as a key to indicate that this is a legacy configuration
+	ymlConfig = append([]yaml.MapItem{{Key: "legacy-config", Value: true}}, ymlConfig...)
+
+	ymlCfgBytes, err := yaml.Marshal(ymlConfig)
+	if err != nil {
+		return errors.Wrapf(err, "failed to marshal YAML")
+	}
+	upgradedCfgBytes, err := upgradeTask.Run(ymlCfgBytes, global, stdout)
+	if err != nil {
+		return errors.Wrapf(err, "failed to upgrade configuration")
+	}
+
+	// back up old configuration
+	if err := backupConfigFile(legacyConfigFilePath, dryRun, stdout); err != nil {
+		return errors.Wrapf(err, "failed to back up legacy configuration file")
+	}
+
+	// back up destination file if it already exists
+	dstFilePath := path.Join(configDirPath, upgradeTask.ConfigFile)
+	if err := backupConfigFile(dstFilePath, dryRun, stdout); err != nil {
+		return errors.Wrapf(err, "failed to back up existing configuration file")
+	}
+
+	// upgraded configuration is empty: no need to write
+	if string(upgradedCfgBytes) == "" {
+		return nil
+	}
+
+	if !dryRun {
+		// write migrated configuration
+		if err := ioutil.WriteFile(dstFilePath, upgradedCfgBytes, 0644); err != nil {
+			return errors.Wrapf(err, "failed to write upgraded configuration")
+		}
+	}
+	printUpgradedConfig(upgradeTask.ConfigFile, upgradedCfgBytes, dryRun, printContent, stdout)
+	return nil
+}
+
+func processUnhandledYMLFiles(configDir string, unknownYMLFiles []string, dryRun bool, stdout io.Writer) error {
+	if len(unknownYMLFiles) == 0 {
+		return nil
+	}
+
+	var unknownNonEmptyFiles []string
+	for _, currUnknownFile := range unknownYMLFiles {
+		currPath := path.Join(configDir, currUnknownFile)
+		bytes, err := ioutil.ReadFile(currPath)
+		if err != nil {
+			return errors.Wrapf(err, "failed to read configuration file")
+		}
+		// if unknown file is empty, just back it up
+		if string(bytes) == "" {
+			if err := backupConfigFile(currPath, dryRun, stdout); err != nil {
+				return err
+			}
+			continue
+		}
+		unknownNonEmptyFiles = append(unknownNonEmptyFiles, currUnknownFile)
+	}
+
+	if len(unknownNonEmptyFiles) == 0 {
+		return nil
+	}
+
+	// if non-empty unknown files were present, print warning
+	dryRunPrintln(stdout, dryRun, fmt.Sprintf(`WARNING: The following configuration file(s) were non-empty and had no known upgraders for legacy configuration: %v`, unknownNonEmptyFiles))
+	dryRunPrintln(stdout, dryRun, fmt.Sprintf(`         If these configuration file(s) are for plugins, add the plugins to the configuration and rerun the upgrade task.`))
+	return nil
 }
